improved logging

+++ aircraft-metadata.lrplugin/Utilities.lua
@@ -27,63 +27,72 @@ LrHttp = import 'LrHttp'
 LrProgressScope = import 'LrProgressScope'
 LrErrors = import 'LrErrors'
 LrStringUtils = import 'LrStringUtils'
+LrSystemInfo = import 'LrSystemInfo'
+LrPrefs = import 'LrPrefs'.prefsForPlugin()
+LrLogger = import 'LrLogger'('AircraftMetadata')
 
-prefs = import 'LrPrefs'.prefsForPlugin()
-logger = import 'LrLogger'('AircraftMetadata')
-
-------- clearLogfile() --------------------------------------------------------
--- clear old logfile
-function clearLogfile()
-	local logPath = LrPathUtils.child(LrPathUtils.getStandardFilePath('documents'), 'AircraftMetadata.log')
+------- startLogger() --------------------------------------------------------
+-- start logger
+function startLogger(functionName)
+	-- check if logging enabled
+	if LrPrefs.prefFlagLogging then
+		LrLogger:enable('logfile')
+		-- clear old logfile
+		logPath = LrPathUtils.child(LrPathUtils.getStandardFilePath('documents'), 'AircraftMetadata.log')
 		if LrFileUtils.exists( logPath ) then
 			success, reason = LrFileUtils.delete( logPath )
 			if not success then
-				logger:error('error deleting existing logfile!'..reason)
+				LrLogger:error('error deleting existing logfile!'..reason)
 			end
+		end
+	else
+		LrLogger:disable()
 	end
+	LrLogger:info('>>>> running '..functionName)
+	LrLogger:info('Lightroom version: '..LrApplication.versionString()..' on '..LrSystemInfo.summaryString())
 end
 
 ------- loadPrefs() -----------------------------------------------------------
 -- load saved preferences
 function loadPrefs()
 	-- lookup KeywordRegNotFound
-	if (prefs.prefKeywordRegNotFound == nil or prefs.prefKeywordRegNotFound == '') then
+	if (LrPrefs.prefKeywordRegNotFound == nil or LrPrefs.prefKeywordRegNotFound == '') then
 		LrErrors.throwUserError('Please set KeywordRegNotFound')
 	end
 	-- lookup URL
-	if (prefs.prefLookupUrl == nil or prefs.prefLookupUrl == '') then
+	if (LrPrefs.prefLookupUrl == nil or LrPrefs.prefLookupUrl == '') then
 		LrErrors.throwUserError('Please set URL for lookup')
 	end
 	-- lookup RegistrationToken1
-	if (prefs.prefRegistrationToken1 == nil or prefs.prefRegistrationToken1 == '') then
+	if (LrPrefs.prefRegistrationToken1 == nil or LrPrefs.prefRegistrationToken1 == '') then
 		LrErrors.throwUserError('Please set registration token 1')
 	end
 	-- lookup RegistrationToken2
-	if (prefs.prefRegistrationToken2 == nil or prefs.prefRegistrationToken2 == '') then
+	if (LrPrefs.prefRegistrationToken2 == nil or LrPrefs.prefRegistrationToken2 == '') then
 		LrErrors.throwUserError('Please set registration token 2')
 	end
 	-- lookup AirlineToken1
-	if (prefs.prefAirlineToken1 == nil or prefs.prefAirlineToken1 == '') then
+	if (LrPrefs.prefAirlineToken1 == nil or LrPrefs.prefAirlineToken1 == '') then
 		LrErrors.throwUserError('Please set airline token 1')
 	end
 	-- lookup AirlineToken2
-	if (prefs.prefAirlineToken2 == nil or prefs.prefAirlineToken2 == '') then
+	if (LrPrefs.prefAirlineToken2 == nil or LrPrefs.prefAirlineToken2 == '') then
 		LrErrors.throwUserError('Please set airline token 2')
 	end
 	-- lookup AircraftToken1
-	if (prefs.prefAircraftToken1 == nil or prefs.prefAircraftToken1 == '') then
+	if (LrPrefs.prefAircraftToken1 == nil or LrPrefs.prefAircraftToken1 == '') then
 		LrErrors.throwUserError('Please set aircraft token 1')
 	end
 	-- lookup AircraftToken2
-	if (prefs.prefAircraftToken2 == nil or prefs.prefAircraftToken2 == '') then
+	if (LrPrefs.prefAircraftToken2 == nil or LrPrefs.prefAircraftToken2 == '') then
 		LrErrors.throwUserError('Please set aircraft token 2')
 	end
 	-- lookup ManufacturerToken1
-	if (prefs.prefManufacturerToken1 == nil or prefs.prefManufacturerToken1 == '') then
+	if (LrPrefs.prefManufacturerToken1 == nil or LrPrefs.prefManufacturerToken1 == '') then
 		LrErrors.throwUserError('Please set manufacturer token 1')
 	end
 	-- lookup ManufacturerToken2
-	if (prefs.prefManufacturerToken2 == nil or prefs.prefManufacturerToken2 == '') then
+	if (LrPrefs.prefManufacturerToken2 == nil or LrPrefs.prefManufacturerToken2 == '') then
 		LrErrors.throwUserError('Please set manufacturer token 2')
 	end
 end
